add routes and app container
