+++ src/containers/app/selectors.js
@@ -1,0 +1,3 @@
+const propsSelector = ({ auth: { isLoggedIn } }) => ({ isLoggedIn });
+
+export default { propsSelector };
